Remove scorecard from workflows
The scorecard workflow is not meant to be executed as part of a pull
request. Instead, it executes on a schedule or when certain changes are
done to a repository. The required rulesets feature is meant for things
that should happen before a merge. The scorecard workflow clearly falls
out of that scope.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>